Move events out of task and create graph_path.
Move FL_BehavioralEvents to a top-level column of Epochs, rather than a component of FL_BehavioralTask. We probably want to keep FL_BehavioralTask as a standalone/metadata descripition of a task/apparatus, not including data of a particular instance of that task.

Create another top-level container called FL_BehavioralGraphPath to describe an animal's path through an FL_BehavioralGraph in an epoch.