FrankLab: Make TimeIntervals iterable.
Use this functionality in the spatial spiking plot.
Currently the iterator returns an 2x2 numpy array with the atomic interval bounds. We might want to refactor this to 1D array (i.e. just returning the atomic interval bounds), but then we will need to allow TimeIntervals to be instantiated with atomic intervals or 1D arrays.